reorganize dataflow plugin
