+++ components/bpmn-q/modeler-component/extensions/dataflow/framework-config/DataConfigManager.js
