+++ components/bpmn-q/modeler-component/extensions/dataflow/modeling/properties-panel/KeyValueEntry.js
